add machine-translated English manual
This commit adds a mechanical translation of the manual from
Russian to English. Manual corrections were made to formatting,
code samples, and the Markdown syntax. No edits were made to the
translated English prose, and no code samples were tested.

Our plan is to use this as a basis for a proper English
translation of the STEAD3 manual.